+++ NEWS.md
@@ -1,6 +1,17 @@
 tag next
 --------------------------------
 
+tag k0.8.2
+--------------------------------
+- ffmpeg: fix some bug
+- ijkio: update and modify features
+- ijkplayer: support don't calculate real frame rate, the first frame will speed up
+
+tag k0.8.1
+--------------------------------
+- ffmpeg: support dns cache
+- ijkio: support inject extra node
+
 tag k0.8.0
 --------------------------------
 - ffmpeg: upgrade to 3.3